11 ways to wear a white plain t-shirt
A white t-shirt is a must-have wardrobe essential. White color is universal and it gets along with any color, so, is there anything more classic than a white tee? Simple, clean, and easy. It literally goes with everything, and never feels undone. It really is the perfect wardrobe staple. Have you ever noticed just how often we stuff our clothes with hard-to-match items? We keep clothes that don’t fit and don’t flatter, simply because we can’t bring ourselves to let them go.

https://smashmart.in/