Inevitable ideas to style a solid polo t-shirt
Solid polo t-shirt is also known as a golf t-shirt as it was first used for playing polo, but now it has crossed the barrier from simply being a uniform to a style staple. Men and women both are in love with this collared t-shirt as it can be dressed up or down on any occasion. The way they look stylish on everyone regardless of any body type, color. It is considered a smart casual. It works wonders in layers. You can layer it in million ways like a blazer, shirt, bomber jacket, etc.

So, get amazing outfit ideas at https://smashmart.in/