How to slay in a pink t-shirt?
A pink t-shirt as an outfit for ladies is a misconception of many men. Many of them, still find it challenging to rock in this color, but that doesn’t mean you shouldn’t try this. Pink looks good on men if styled in the right way. Pink can make your wardrobe a classy one. Men look seriously stylish in pink because it flatters every skin tone due to the many shades available. 74.3% of women find men attractive in pink color. So, it’s worth trying it. 

So, get amazing outfit ideas at https://smashmart.in/