How to rock a pink shirt?
A pink shirt is something which most of us don’t have in our wardrobe, but we should consider adding it. It might sound daunting to wear pink, but it can work wonders if paired well. While the idea that the pink color is only for women is outdated. It’s time for you to get rid of outdated minds and move into the 21st century. A gentleman doesn’t hesitate to carry pink as wearing pink shows that you are confident and calm. It makes your work easier to charm ladies by your looks. Most of us have the misconception that “Wearing pink is effeminate” Color has no limits, it’s all about how you style it.

So, get amazing outfit ideas at https://classicpolos.com/