How to wear Blue shirt in a right way?
Wear a blue shirt in the right way with the help of this guide. The key to get the blue shirt right is to pair it with the right garments. You can spruce up the outfit in both formal and casual ways. Most people underestimate a classic blue shirt as a boring shirt, but if you trust us it is the most versatile piece you can own. To be honest, you shouldn’t treat a blue shirt as a boring basic. Every man looks good in a basic blue shirt regardless of body shape and color. 

Get amazing outfit ideas at https://classicpolos.com/